chore(localize): split number related functions into their own files
