+++ packages/localize/index.js
@@ -3,13 +3,11 @@ export { getDateFormatBasedOnLocale } from './src/date/getDateFormatBasedOnLocal
 export { getMonthNames } from './src/date/getMonthNames.js';
 export { getWeekdayNames } from './src/date/getWeekdayNames.js';
 export { parseDate } from './src/date/parseDate.js';
-export {
-  formatNumber,
-  formatNumberToParts,
-  getFractionDigits,
-  getDecimalSeparator,
-  getGroupSeparator,
-} from './src/formatNumber.js';
+export { formatNumber } from './src/number/formatNumber.js';
+export { formatNumberToParts } from './src/number/formatNumberToParts.js';
+export { getDecimalSeparator } from './src/number/getDecimalSeparator.js';
+export { getFractionDigits } from './src/number/getFractionDigits.js';
+export { getGroupSeparator } from './src/number/getGroupSeparator.js';
 export { localize, setLocalize } from './src/localize.js';
 export { LocalizeManager } from './src/LocalizeManager.js';
 export { LocalizeMixin } from './src/LocalizeMixin.js';
