+++ packages/localize/test/number/getFractionDigits.test.js
@@ -1,0 +1,11 @@
+import { expect } from '@open-wc/testing';
+
+import { getFractionDigits } from '../../src/number/getFractionDigits.js';
+
+describe('getFractionDigits', () => {
+  it('returns number of fraction digits for currency', async () => {
+    expect(getFractionDigits('JPY')).to.equal(0);
+    expect(getFractionDigits('EUR')).to.equal(2);
+    expect(getFractionDigits('BHD')).to.equal(3);
+  });
+});
